Fixed mmrm function

+++ micromamba.zsh
@@ -7,11 +7,15 @@ alias mmc='micromamba create'
 alias mmls='ls $MAMBA_ROOT_PREFIX/envs'
 
 mmrm() {
-  if [[ "$1" != "" ]]; then
-    echo "Deleting environment $1"
-    rm -rf "$MAMBA_ROOT_PREFIX"/envs/"$1"
+  if [[ -n "$MAMBA_ROOT_PREFIX" ]]; then
+    if [[ "$1" != "" ]]; then
+      echo "Deleting environment $1"
+      rm -rf "$MAMBA_ROOT_PREFIX"/envs/"$1"
+    else
+      echo "Environment name parameter needed!"
+    fi
   else
-    echo "Environment name parameter needed!"
+    echo "MAMBA_ROOT_PREFIX variable is not set!"
   fi
 }
